Rename files to adapt to suggest header extension and R files extension
