+++ R/functions_estimation_engine_c.R
