+++ src/Makevars.in
@@ -1,4 +1,4 @@
 ## -*- mode: makefile; -*-
 
 PKG_CXXFLAGS =  $(SHLIB_OPENMP_CXXFLAGS)
-PKG_LIBS= $(SHLIB_OPENMP_CFLAGS) $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
+PKG_LIBS= $(SHLIB_OPENMP_CXXFLAGS) $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
