+++ src/gather_receiver_model.cpp
@@ -2,7 +2,7 @@
 // [[Rcpp::depends(RcppArmadillo)]]
 using namespace Rcpp;
 using namespace arma;
-#include "gather_progress.hpp"
+#include "gather_progress.h"
 
 // @inherit estimate_REM params
